Math Experimentation Completed

+++ Math/Quadratic.java
@@ -14,19 +14,27 @@ public class Quadratic
 	 */
 	public static void main(String[] args)
 	{
-		int a, b, c;  // ax^2 + bx + c
+		double a, b, c;  // ax^2 + bx + c
 		double discriminant, root1, root2;
 
 		Scanner scan = new Scanner(System.in);
 
 		System.out.print("Enter the coefficient of x squared: ");
-		a = scan.nextInt();
+		a = (int) scan.nextDouble();
 
 		System.out.print("Enter the coefficient of x: ");
-		b = scan.nextInt();
+		b = (int) scan.nextDouble();
 
 		System.out.print("Enter the constant: ");
-		c = scan.nextInt();
+		c = (int) scan.nextDouble();
+
+		// 2, 16, and 3 results came out to be Root #1 = -0.19211344706804567 and Root #2 = -7.807886552931954 //
+
+		// 2.5, 16.8, and 3.2 results came out to be Root #1 =  -0.19211344706804567, and Root #2 =  -7.807886552931954 //
+
+		// Results came out to be the same because of how the numbers are similar to each other. I also think the reason why the results are the same is because of both codes are still a double instead of being an integer. //
+
+		// I'm not sure what happened after I removed the cast and changed the declaration for a, b, and c. The results came out to be very complicated and I wasn't able to enter numbers anymore. //
 
 		// Use the quadratic formula to compute the roots.
 		// Assumes a positive discriminant. (otherwise we need complex numbers)
@@ -39,5 +47,6 @@ public class Quadratic
 		System.out.println("Root #2: " + root2);
 
 		scan.close();
+
 	}
 }
